Release 1.4
After running for several months, the modem sometimes gets into an unresponsive state. Rebooting the Raspberry Pico did not reset the modem in earlier versions, so fixing this required a power cycle.

In Release 1.4, the modem gets power cycled when the Raspberry Pico boots. The Pico also checks regularly (by default, every 24 hours) whether the modem is responsive, and reboots if it is not found responsive.